pyproject.toml
[project]
name = "ig-agent"
version = "0.1.0"
description = "Autonomous niche-targeted Instagram content pipeline"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Instagram
    "instagrapi>=2.1.3",
    "instaloader>=4.13",

    # LLM clients (OpenAI-compat across providers)
    "openai>=1.54",
    "httpx>=0.27",
    "tenacity>=9.0",

    # Config + validation
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # CLI
    "typer>=0.15",
    "rich>=13.9",
    "questionary>=2.0",

    # Scheduling
    "apscheduler>=3.10",

    # Image/video/audio
    "pillow>=11.0",
    "imagehash>=4.3",
    "numpy>=2.0",
    "moviepy>=2.1",
    "playwright>=1.48",
    "edge-tts>=7.0",
    "faster-whisper>=1.1",

    # Scraping + stock
    "requests>=2.32",
    "beautifulsoup4>=4.12",

    # Challenge resolver
    "pyotp>=2.9",
    "imap-tools>=1.7",

    # Alerts
    "aiohttp>=3.10",

    # Dashboard
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "jinja2>=3.1",

    # Object storage (optional; only used when R2 configured)
    "boto3>=1.35",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
# WhisperX gives better word-level alignment (wav2vec2 forced align) but
# pulls in a heavy torchaudio stack. Optional — faster-whisper remains the
# default transcriber when whisperx isn't installed.
whisperx = [
    "whisperx>=3.1",
]
# Finish pass — upscale (Real-ESRGAN BSD-3) + face restore (GFPGAN Apache-2.0).
# Heavy torch stack; optional. CodeFormer (S-Lab NC) deliberately excluded.
finish = [
    "realesrgan>=0.3",
    "gfpgan>=1.3",
    "basicsr>=1.4",
    "opencv-python-headless>=4.10",
]
# HF Space fallback for the finish pass (no local GPU).
finish-hf = [
    "gradio-client>=1.0",
]
# Niche RAG — local embeddings fallback + PDF ingestion. Commercial-safe.
# Without these, RAG still works using Gemini free-tier embeddings and
# .txt/.md sources.
rag = [
    "sentence-transformers>=2.7",
    "pypdf>=4.0",
]
# Beat-synced reel cuts. librosa is ISC — commercial-OK. Pulls numba.
# We deliberately avoid BeatNet and beat_this (both CC-BY-NC).
beat = [
    "librosa>=0.10",
]
# Local aesthetic scoring — cheap first-pass rank before the vision
# LLM runs. This extra is COMMERCIAL-SAFE (LAION aesthetic predictor is MIT).
aesthetic = [
    "simple-aesthetics-predictor>=0.1.2",   # LAION CLIP predictor (MIT)
    "transformers>=4.40",
    "torch>=2.0",
]
# OPT-IN NON-COMMERCIAL quality lane. pyiqa ships under PolyForm
# Noncommercial — fine for personal / research runs, NOT safe for a
# monetised Instagram page. Users who accept this install it themselves;
# the default `[aesthetic]` never pulls it.
aesthetic-nc = [
    "pyiqa>=0.1.10",                        # PolyForm Noncommercial
]
# Reddit question harvester — PRAW is BSD-2 (commercial-safe). Free API
# tier requires a script-app client_id + client_secret + user_agent.
reddit = [
    "praw>=7.7",
]
# Stable Audio Open Small — generative music. Weights are Stability AI
# Community Licence (free up to USD 1M annual revenue; Enterprise above).
# The user must acknowledge the licence in niche.yaml before the model
# runs. Pulls torch + stable-audio-tools; large install (~3 GB).
stable-audio = [
    "stable-audio-tools>=0.0.19",
    "torch>=2.0",
    "torchaudio>=2.0",
]

[project.scripts]
ig-agent = "src.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
